# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Xsolla Marketplace API",
        "description": "This is a markeplace for purchasing and selling game keys.",
        "contact": {
            "url": "https://github.com/InfluxOW",
            "email": "krochak_n@mail.ru"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/games": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Games Index",
                "description": "View all games",
                "operationId": "gamesIndex",
                "parameters": [
                    {
                        "name": "filter[platform]",
                        "in": "query",
                        "description": "Filter games by platform:slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[distributor]",
                        "in": "query",
                        "description": "Filter games by distributor:slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[available]",
                        "in": "query",
                        "description": "Leave only games that are available for purchase",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "filter[available_at]",
                        "in": "query",
                        "description": "Leave only games that are available for purchase at the specified distributor:slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[price_lte]",
                        "in": "query",
                        "description": "Leave only games that costs less or equals to the specified price",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filter[price_gte]",
                        "in": "query",
                        "description": "Leave only games that costs more or equals to the specified price",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort games by one of the available params: price, name or platform. Default sort direction is ASC. To apply DESC sort add '-' symbol before the param name.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Results page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Games were fetched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "games": {
                                            "properties": {
                                                "0": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object",
                                                        "$ref": "#/components/schemas/GameResource"
                                                    },
                                                    "collectionFormat": "multi"
                                                }
                                            },
                                            "type": "object",
                                            "collectionFormat": "multi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Games Store",
                "description": "Store a new game",
                "operationId": "gamesStore",
                "requestBody": {
                    "description": "Game data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "platform"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "The Witcher 3: Wild Hunt"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "The Witcher 3: Wild Hunt is a 2015 action role-playing game developed and published by Polish developer CD Projekt Red and is based on The Witcher series of fantasy novels by Andrzej Sapkowski. It is the sequel to the 2011 game The Witcher 2: Assassins of Kings and the third main installment in the The Witcher's video game series, played in an open world with a third-person perspective."
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "50"
                                    },
                                    "platform": {
                                        "type": "string",
                                        "example": "PC"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Game has been stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "object",
                                        "$ref": "#/components/schemas/GameResource"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Game has not been stored due to validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": [
                                                            "Name attribute is required."
                                                        ]
                                                    },
                                                    "collectionFormat": "multi"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "You should be authorized to access the endpoint"
                    },
                    "403": {
                        "description": "You don't have permissions to access the endpoint"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/games/{game:slug}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Games Show",
                "description": "View specified game",
                "operationId": "gamesShow",
                "parameters": [
                    {
                        "name": "game:slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Specified game has been fetched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "object",
                                        "$ref": "#/components/schemas/GameResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Specified game has not been found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "GameResource": {
                "properties": {
                    "name": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "The Witcher 3: Wild Hunt"
                    },
                    "description": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "The Witcher 3: Wild Hunt is a 2015 action role-playing game developed and published by Polish developer CD Projekt Red and is based on The Witcher series of fantasy novels by Andrzej Sapkowski. It is the sequel to the 2011 game The Witcher 2: Assassins of Kings and the third main installment in the The Witcher's video game series, played in an open world with a third-person perspective."
                    },
                    "platform": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "PC"
                    },
                    "price": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "50"
                    },
                    "link": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "http://localhost:8000/api/games/the-witcher-3-wild-hunt"
                    },
                    "available_keys": {
                        "type": "string[]",
                        "readOnly": "true",
                        "example": {
                            "Uplay": 3,
                            "Steam": 2
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "GameResource"
                }
            },
...
        },
        "securitySchemes": {
            "access_token": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}